+++ dflib/src/main/java/com/nhl/dflib/aggregate/SeriesMinMax.java
@@ -7,6 +7,60 @@ import com.nhl.dflib.Series;
  */
 public class SeriesMinMax {
 
+    public static <T extends Comparable<T>> T min(Series<? extends T> s) {
+
+        int size = s.size();
+        if (size == 0) {
+            return null;
+        }
+
+        T min = s.get(0);
+
+        int i = 1;
+        for (; min == null && i < size; i++) {
+            min = s.get(i);
+        }
+
+        for (; i < size; i++) {
+
+            T t = s.get(i);
+            if (t != null) {
+                if (t.compareTo(min) < 0) {
+                    min = t;
+                }
+            }
+        }
+
+        return min;
+    }
+
+    public static <T extends Comparable<T>> T max(Series<? extends T> s) {
+
+        int size = s.size();
+        if (size == 0) {
+            return null;
+        }
+
+        T max = s.get(0);
+
+        int i = 1;
+        for (; max == null && i < size; i++) {
+            max = s.get(i);
+        }
+
+        for (; i < size; i++) {
+
+            T t = s.get(i);
+            if (t != null) {
+                if (max.compareTo(t) < 0) {
+                    max = t;
+                }
+            }
+        }
+
+        return max;
+    }
+
     public static int maxInt(Series<? extends Number> s) {
 
         int size = s.size();